mfenc: Await extradata at the end of the init function, if not available immediately
The Qualcomm H264 encoder doesn't provide extradata immediately,
but a short while after starting, the extradata is available (but
I haven't seen any signal that one can wait for).

The Qualcomm HEVC encoder, on the other hand, doesn't seem to export
extradata (MF_MT_MPEG_SEQUENCE_HEADER) at all - that can be an issue
for use with muxers that require it.