WIP: arm: Add helpers for debug printing from assembly
Use e.g. as
    DUMP_SIMD_S16("input", 2, 5)
which then prints

    634: input:
    v02 = { 14468, -1, 27728, -1, 14468, -1, 27728, -1 }
    v03 = { 17472, -1, 29952, -1, 17472, -1, 29952, -1 }
    v04 = { 2002, 3024, 1637, 1702, 0, 0, 0, 0 }
    v05 = { 5026, 3339, 5026, 3339, 0, 0, 0, 0 }

where 634 is the line number of the printout, "input" is a generic
label for what is printed, and for a call to DUMP_SIMD_xx(name, start, end),
it prints the range of registers from start to end, inclusively.

There's various ways of interpreting SIMD registers, and there's
also a similar DUMP_GPR() for printing the contents of GPRs.