aarch64: fft: Simplify loading of large constants
Use the ldr pseudo-instruction, which in practice dumps a (in this
case) 32 bit constant nearby (e.g. at the end of the current section)
and loads it with a pc-relative load. Using the w register names,
to only store a 32 bit constant (where loading it will implicitly
clear the upper half).