checkasm: allow run-time fallback to AV_READ_TIME
On Linux RISC-V, depending on kernel version and configuration, the
cycle counters may be available:
- via Linux perf (as on Arm),
- directly with RDCYCLE (which is considered legacy),
- neither (if admistratively disabled), or
- or both (in legacy compatibility mode).

This allows to try Linux perf, with a fallback to RDCYCLE. Without
this, benchmarking would breakdown in upcoming Linux RISC-V kernel
releases. This also paves the way for falling back from the cycle
counter on Arm or RISC-V, whence they too are disabled.

Signed-off-by: Martin Storsjö <martin@martin.st>