checkasm: Generalize crash handling
This replaces the riscv specific handling from
7212466e735aa187d82f51dadbce957fe3da77f0 (which essentially is
reverted, together with 286d6742218ba0235c32876b50bf593cb1986353)
with a more generic version imported from dav1d's checkasm,
supporting both Unix and Windows.

See in particular dav1d commits
0b6ee30eab2400e4f85b735ad29a68a842c34e21 and
0421f787ea592fd2cc74c887f20b8dc31393788b, authored by
Henrik Gramner.

The overall approach is the same; set up a signal handler,
store the state with setjmp/sigsetjmp, jump out of the crashing
function with longjmp/siglongjmp.

The main, sublte difference, is in what happens when the signal
handler is invoked. In the previous implementation, it would
resume from right before calling the crashing function, and
then skip that call. In the imported implementation from dav1d,
we jump back to the call to declare_func(), which precedes
check_func(). When resuming after the signal handler, we return
to the start of the function; when we hit calls to check_func()
with the same function pointer that was already tested, that
test gets skipped and we proceed to whatever was next to be
tested.

Other differences are:
- Support for other signal handling mechanisms (Windows
  AddVectoredExceptionHandler)
- Using RtlCaptureContext/RtlRestoreContext instead of setjmp/longjmp
  on Windows with SEH
- Only catching signals once per function
- Not using an arch specific signal handler written in assembly

Using a dummy av_unused variable declaration in declare_func() to
contain the call to checkasm_context_save() without making it a
statement; otherwise functions that declare further variables
after declare_func() would get warnings about declarations after
statement.