WIP: Try building for arm too, with MSVC and llvm-mingw
For llvm-mingw, we don't want to use --cross-prefix here as that
enables cross compilation - we can execute the binaries just fine here.

Alternatively we could just use an armv7 native toolchain.

For MSVC, we need to manually disable vfwcap. The DLLs for vfw32
(avicap32.dll) don't exist in 32 bit ARM mode, even if the import
libraries for them do exist in the Windows SDK.