configure: use proper Windows-style static library naming
On Windows, static libraries are typically named with a .lib extension.
An exception to this is MinGW targets, which are treated as a distinct
target-os.

Using Windows-style naming allows Clang to be used as the linker driver,
instead of invoking link or lld-link directly. The latter approach
requires manually specifying standard libraries, which may be
error-prone or incomplete.

This change also improves support for LTO and sanitizer builds, where
it's significantly easier to let the compiler driver manage the
necessary linker flags.

It fixes issues where Clang is asked to `-lavcodec`, which gets passed
to the linker as avcodec.lib, resulting in an error like:
lld-link: error: could not open 'avcodec.lib': no such file or directory
This happens because `libavcodec.a` were unexpectedly generated, not
`avcodec.lib` expected by tooling.

Note that this doesn't affect mingw build. MSVC builds work the same,
because there libraries are passed as files directly to linker command.

This also removes LD_LIB from Win32/64 target as there is one type of
.lib in practice. We cannot build both shared and static as noted by the
next line.

Signed-off-by: Kacper Michajłow <kasper93@gmail.com>